versions_dep/v3/apps/orchestrator/pyproject.toml
[project]
name = "gpthub-orchestrator"
version = "0.1.0"
description = "GPTHub Workspace orchestrator"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Aleksandr Mordvinov" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/FUYOH666"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gpthub_orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
